OneScan/OneScan-main/OneScan-main/requirements.txt
flask==3.0.2
werkzeug==3.0.1
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
requests==2.31.0
pillow==10.2.0
mysql-connector-python==8.2.0
 